refactor: remove brokerage.simple duplicates
